[squash] Rename styles
